LayoutTests/fast/dom/Document/document-charset-expected.txt

Initial

 document.charset: KOI8-R

 document.defaultCharset: defined

 document.characterSet: KOI8-R

 document.inputEncoding: KOI8-R

Setting charset to UTF-8 (expected to fail as it is read-only)

 document.charset: KOI8-R

 document.defaultCharset: defined

 document.characterSet: KOI8-R

 document.inputEncoding: KOI8-R

Setting characterSet to UTF-8 (expected to fail as it is read-only)

 document.charset: KOI8-R

 document.defaultCharset: defined

 document.characterSet: KOI8-R

 document.inputEncoding: KOI8-R
